requirements.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.100.1
orjson==3.9.10
uvicorn==0.27.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
pydantic==2.1.1
pydantic-settings==2.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
